[Core] Grouped Tests
