+++ selcukes-core/src/test/java/io/github/selcukes/core/tests/web/BaseTest.java
@@ -1,0 +1,47 @@
+/*
+ *  Copyright (c) Ramesh Babu Prudhvi.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package io.github.selcukes.core.tests.web;
+
+import io.github.selcukes.commons.Await;
+import io.github.selcukes.core.listener.EventCapture;
+import io.github.selcukes.core.page.WebPage;
+import io.github.selcukes.wdb.driver.LocalDriver;
+import io.github.selcukes.wdb.enums.DriverType;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.events.EventFiringDecorator;
+import org.openqa.selenium.support.events.WebDriverListener;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+public class BaseTest {
+    WebDriver driver;
+    WebPage page;
+
+    @BeforeMethod
+    public void setup() {
+        driver = new LocalDriver().createWebDriver(DriverType.CHROME, true);
+        WebDriverListener eventCapture = new EventCapture();
+        driver = new EventFiringDecorator(eventCapture).decorate(driver);
+        page = new WebPage(driver);
+    }
+
+    @AfterMethod
+    public void teardown() {
+        Await.until(3);
+        driver.quit();
+    }
+}
