+++ selcukes-core/src/test/java/io/github/selcukes/core/tests/web/EventDriverTest.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package io.github.selcukes.core.tests;
+package io.github.selcukes.core.tests.web;
 
 import io.github.selcukes.commons.helper.DateHelper;
 import io.github.selcukes.commons.helper.FileHelper;
